+++ src/main/java/com/trinet/connecto/service/CommentService.java
@@ -5,9 +5,9 @@ import com.trinet.connecto.model.Comment;
 import java.util.List;
 
 public interface CommentService {
-    List<Comment> getAllCommentsForThread(Integer threadId);
+    List<Comment> getAllCommentsForThread(Long threadId);
 
-    Comment getCommentById(Integer commentId);
+    Comment getCommentById(Long commentId);
 
     Comment addNewComment(Comment comment);
 }
